requirements.txt
python-dotenv~=0.19.0
selenium==4.1.0